+++ src/assets/obj_loader.hpp
@@ -9,4 +9,7 @@
 // this one loads in indices as well.
 Model LoadOBJModelButBetter(const std::filesystem::path& path);
 
+// Loads an .obj file but with the actual contents of the file as a parameter.
+Model LoadOBJModelFromString(const std::string stringContents);
+
 #endif